requirements.txt
streamlit==1.35.0
pandas==2.0.2
fastapi==0.112.0
uvicorn==0.30.6
mysql-connector-python==8.0.33
requests==2.31.0
pytest==8.3.2
